Extract EntryOrderConfig
