+++ src/lib/template/seaweed/entry_order_config/EntryOrderConfig.ts
@@ -1,0 +1,3 @@
+export const removeProxyWrapperOnString = (wrapped: string): string => {
+	return wrapped.slice("Proxy<".length, wrapped.length - 1);
+};
